refactor(plugin-crx): merge plugin-crx-bridge into plugin-crx
The two plugins were tightly coupled — the bridge already declared a
dependency on plugin-crx's settings capability, and only composer-app
consumed both. Collapsing them removes a split that wasn't pulling its
weight: the receiver and its settings now live in one plugin with one
activation, one set of translations, and one PLUGIN.mdl.
