--- packages/plugins/plugin-crx-bridge/PLUGIN.mdl
@@ -1,263 +1,0 @@
----
-id: org.dxos.plugin.crxBridge
-name: CrxBridgePlugin
-version: 0.1.0
----
-
-Receives clippings sent from the Composer browser extension (`@dxos/composer-crx`) and
-materializes them as `Person` or `Organization` objects in the currently active space.
-The extension discovers this tab via `chrome.tabs.query` over a configured list of URLs
-and delivers the payload as a same-origin `window` `CustomEvent`.
-
-The extension owns the DOM picker and bridge sender. This plugin owns only the receiver
-side: envelope validation, clip → ECHO object mapping, and user feedback (toast).
-
-## Extensions
-
-The following extension dialects are used in this document.
-Each extension is defined in the Appendix or resolved via its URI.
-
-| Term   | URI                            |
-|--------|--------------------------------|
-| `type` | `org.dxos.mdl.type@1.0`        |
-| `feat` | `org.dxos.mdl.feat@1.0`        |
-| `test` | `org.dxos.mdl.test@1.0`        |
-
-## Types
-
-```mdl
-type ClipKind
-  supported: person | organization
-  note: |
-    At runtime, `Clip.kind` is modeled as `string` so new kinds can ship
-    from the extension without breaking envelope decoding. The receiver
-    accepts `person` and `organization` only; any other value is rejected
-    with `unsupportedKind`.
-```
-
-```mdl
-type ClipSource
-  fields:
-    url: string
-    title: string
-    favicon?: string
-    clippedAt: string           # ISO timestamp
-```
-
-```mdl
-type ClipSelection
-  fields:
-    text: string                # innerText of the picked subtree
-    html?: string               # outerHTML, possibly truncated
-    htmlTruncated?: boolean
-    rect?: { x: number, y: number, width: number, height: number }
-```
-
-```mdl
-type ClipHints
-  desc: |
-    Cheap signals harvested from the whole page at pick time. MVP uses a subset
-    (og:title, og:image) as fallbacks; a future agent stage may interpret all of them.
-  fields:
-    ogTitle?: string
-    ogDescription?: string
-    ogImage?: string
-    jsonLd?: unknown[]
-    h1?: string
-    firstImage?: string
-```
-
-```mdl
-type Clip
-  desc: |
-    The on-the-wire envelope is intentionally permissive so that older
-    receivers can read newer versions just far enough to respond with
-    `unsupportedVersion`. Version-specific validation runs after the
-    envelope decode succeeds.
-  fields:
-    version: number             # runtime check: only 1 is supported
-    kind: string                # runtime check against ClipKind.supported
-    source: ClipSource
-    selection: ClipSelection
-    hints?: ClipHints
-```
-
-```mdl
-type ClipAck
-  fields:
-    ok: boolean
-    id?: string                 # ID of the created object when ok = true
-    error?: string              # stable error code when ok = false
-```
-
-## Features
-
-```mdl
-feat F-1: Clip Ingestion Listener
-
-  req F-1.1:
-    when: plugin is activated
-    then: a single window event listener for `composer:clip` is registered
-
-  req F-1.2:
-    when: the page unloads or the plugin is disabled
-    then: the listener is removed cleanly
-
-  req F-1.3:
-    when: any payload arrives
-    then: a `composer:clip:ack` CustomEvent is dispatched on `window` carrying ClipAck
-```
-
-```mdl
-feat F-2: Envelope Validation
-
-  req F-2.1:
-    when: a `composer:clip` event is received
-    then: the `detail` is decoded against the Clip schema
-
-  req F-2.2:
-    when: the payload fails schema decoding
-    then: ack is { ok: false, error: 'invalidPayload' } and no object is created
-
-  req F-2.3:
-    when: the payload has a version other than 1
-    then: ack is { ok: false, error: 'unsupportedVersion' }
-```
-
-```mdl
-feat F-3: Clip → Object Mapping
-
-  req F-3.1:
-    when: kind = 'person'
-    then: |
-      a Person is created with best-effort fields:
-        fullName  ← hints.h1 ?? hints.ogTitle ?? firstLine(selection.text)
-        image     ← hints.ogImage
-        notes     ← selection.text (truncated to a reasonable size)
-        urls      ← [{ label: source.title, value: source.url }]
-
-  req F-3.2:
-    when: kind = 'organization'
-    then: |
-      an Organization is created with best-effort fields:
-        name        ← hints.ogTitle ?? hints.h1 ?? firstLine(selection.text)
-        description ← hints.ogDescription
-        image       ← hints.ogImage
-        website     ← source.url
-
-  req F-3.3:
-    when: there is no name/title candidate
-    then: the object is still created; empty name is acceptable
-
-  req F-3.4:
-    when: kind is unknown (future kinds)
-    then: ack is { ok: false, error: 'unsupportedKind' } and no object is created
-```
-
-```mdl
-feat F-4: Space Resolution
-
-  req F-4.1:
-    when: the active space cannot be determined
-    then: no object is created; ack is { ok: false, error: 'noSpace' }
-
-  req F-4.2:
-    when: the active space is resolved
-    then: the object is added via SpaceOperation.AddObject with target = space.db
-```
-
-```mdl
-feat F-5: User Feedback
-
-  req F-5.1:
-    when: an object is created successfully
-    then: a success toast is shown naming the kind and (when present) its label
-
-  req F-5.2:
-    when: creation fails
-    then: an error toast is shown with the stable error code
-```
-
-## Acceptance
-
-```mdl
-test T-1: Valid person clip creates Person
-  given: plugin active, a default space present
-  when: window dispatches `composer:clip` with kind=person, hints.h1='Jane Doe'
-  then:
-    - a Person is created with fullName='Jane Doe'
-    - an ack event carries { ok: true, id: <non-empty> }
-    - a success toast is shown
-```
-
-```mdl
-test T-2: Organization mapping falls back to og:title
-  given: plugin active
-  when: clip arrives with kind=organization, hints.ogTitle='DXOS', hints.h1 undefined
-  then:
-    - Organization.name === 'DXOS'
-    - Organization.website === source.url
-```
-
-```mdl
-test T-3: Invalid payload is rejected
-  given: plugin active
-  when: window dispatches `composer:clip` with a malformed detail (missing kind)
-  then:
-    - no object is created
-    - ack is { ok: false, error: 'invalidPayload' }
-```
-
-```mdl
-test T-4: Missing name is allowed
-  given: plugin active
-  when: clip arrives with kind=person and no hints and empty selection.text
-  then:
-    - a Person is created (empty fullName)
-    - ack is { ok: true, id: <non-empty> }
-```
-
-```mdl
-test T-5: Unsupported envelope version
-  given: plugin active
-  when: clip arrives with version=2
-  then:
-    - no object is created
-    - ack is { ok: false, error: 'unsupportedVersion' }
-```
-
----
-
-## Appendix: Extension Definitions
-
-```mdl
-ext type
-  uri: dxos.spec/ext/type@1.0
-  desc: A named data structure with typed fields and optional literals.
-  fields:
-    desc?: Prose
-    fields?: FieldMap
-    literals?: UnionList
-    extends?: TypeRef[]
-```
-
-```mdl
-ext feat
-  uri: dxos.spec/ext/feat@1.0
-  desc: A named feature grouping one or more requirements.
-  fields:
-    desc?: Prose
-    req: RequirementList
-  nesting: self
-```
-
-```mdl
-ext test
-  uri: dxos.spec/ext/test@1.0
-  desc: An acceptance scenario expressed as given / when / then steps.
-  fields:
-    given?: Step | Step[]
-    when?: Step | Step[]
-    then: Step | Step[]
-    tags?: TagList
-```
