--- packages/plugins/plugin-crx-bridge/moon.yml
@@ -1,11 +1,0 @@
-layer: library
-language: typescript
-tags:
-  - ts-build
-  - ts-test
-tasks:
-  compile:
-    args:
-      - '--entryPoint=src/index.ts'
-      - '--entryPoint=src/translations.ts'
-      - '--entryPoint=src/types/index.ts'
