--- packages/plugins/plugin-crx-bridge/package.json
@@ -1,69 +1,0 @@
-{
-  "name": "@dxos/plugin-crx-bridge",
-  "version": "0.8.3",
-  "private": true,
-  "description": "Receives clippings from the Composer browser extension.",
-  "homepage": "https://dxos.org",
-  "bugs": "https://github.com/dxos/dxos/issues",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/dxos/dxos"
-  },
-  "license": "MIT",
-  "author": "DXOS.org",
-  "sideEffects": true,
-  "type": "module",
-  "imports": {
-    "#capabilities": "./src/capabilities/index.ts",
-    "#meta": "./src/meta.ts",
-    "#translations": "./src/translations.ts",
-    "#types": "./src/types/index.ts"
-  },
-  "exports": {
-    ".": {
-      "source": "./src/index.ts",
-      "types": "./dist/types/src/index.d.ts",
-      "browser": "./dist/lib/browser/index.mjs",
-      "node": "./dist/lib/node-esm/index.mjs"
-    },
-    "./translations": {
-      "source": "./src/translations.ts",
-      "types": "./dist/types/src/translations.d.ts",
-      "browser": "./dist/lib/browser/translations.mjs",
-      "node": "./dist/lib/node-esm/translations.mjs"
-    },
-    "./types": {
-      "source": "./src/types/index.ts",
-      "types": "./dist/types/src/types/index.d.ts",
-      "browser": "./dist/lib/browser/types/index.mjs",
-      "node": "./dist/lib/node-esm/types/index.mjs"
-    }
-  },
-  "types": "dist/types/src/index.d.ts",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "dependencies": {
-    "@dxos/app-framework": "workspace:*",
-    "@dxos/app-toolkit": "workspace:*",
-    "@dxos/echo": "workspace:*",
-    "@dxos/invariant": "workspace:*",
-    "@dxos/log": "workspace:*",
-    "@dxos/operation": "workspace:*",
-    "@dxos/plugin-client": "workspace:*",
-    "@dxos/plugin-markdown": "workspace:*",
-    "@dxos/plugin-space": "workspace:*",
-    "@dxos/react-ui": "workspace:*",
-    "@dxos/types": "workspace:*",
-    "@dxos/util": "workspace:*",
-    "effect": "catalog:"
-  },
-  "devDependencies": {
-    "vite": "catalog:",
-    "vitest": "catalog:"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
