--- packages/plugins/plugin-crx-bridge/src/capabilities/index.ts
@@ -1,8 +1,0 @@
-//
-// Copyright 2026 DXOS.org
-//
-
-// No capabilities are currently contributed lazily; the plugin installs a
-// window listener at startup and surfaces feedback via toasts. Kept as a barrel
-// so future surfaces (e.g., a settings UI or an options page) can be added here.
-export {};
