--- packages/plugins/plugin-crx-bridge/src/index.ts
@@ -1,11 +1,0 @@
-//
-// Copyright 2026 DXOS.org
-//
-
-import { Plugin } from '@dxos/app-framework';
-
-import { meta } from './meta';
-
-export const CrxBridgePlugin = Plugin.lazy(meta, () => import('./CrxBridgePlugin'));
-
-export * from './meta';
