--- packages/plugins/plugin-crx-bridge/src/types/index.ts
@@ -1,5 +1,0 @@
-//
-// Copyright 2026 DXOS.org
-//
-
-export * as Clip from './Clip';
