--- packages/plugins/plugin-crx-bridge/tsconfig.json
@@ -1,57 +1,0 @@
-{
-  "extends": [
-    "../../../tsconfig.base.json"
-  ],
-  "compilerOptions": {
-    "types": [
-      "node"
-    ]
-  },
-  "exclude": [
-    "*.t.ts",
-    "vite.config.ts"
-  ],
-  "include": [
-    "src/**/*.ts",
-    "src/**/*.tsx",
-    "src/*.ts"
-  ],
-  "references": [
-    {
-      "path": "../../common/invariant"
-    },
-    {
-      "path": "../../common/log"
-    },
-    {
-      "path": "../../common/util"
-    },
-    {
-      "path": "../../core/compute/operation"
-    },
-    {
-      "path": "../../core/echo/echo"
-    },
-    {
-      "path": "../../sdk/app-framework"
-    },
-    {
-      "path": "../../sdk/app-toolkit"
-    },
-    {
-      "path": "../../sdk/types"
-    },
-    {
-      "path": "../../ui/react-ui"
-    },
-    {
-      "path": "../plugin-client"
-    },
-    {
-      "path": "../plugin-markdown"
-    },
-    {
-      "path": "../plugin-space"
-    }
-  ]
-}
