+++ packages/plugins/plugin-crx/moon.yml
@@ -2,6 +2,7 @@ layer: library
 language: typescript
 tags:
   - ts-build
+  - ts-test
 tasks:
   compile:
     args:
