+++ packages/plugins/plugin-crx/src/mapping.test.ts
