+++ packages/plugins/plugin-crx/src/mapping.ts
