+++ packages/plugins/plugin-crx/src/meta.ts
@@ -9,8 +9,9 @@ export const meta: Plugin.Meta = {
   id: 'org.dxos.plugin.crx',
   name: 'CRX',
   description: trim`
-    Manages how Composer coordinates with the composer-crx browser extension.
-    Owns the user-facing settings surface for the extension integration.
+    Coordinates with the composer-crx browser extension. Owns the settings
+    surface and receives clippings from the extension, materializing them as
+    Person, Organization, or Note objects in the active space.
   `,
   icon: 'ph--browser--regular',
   iconHue: 'neutral',
