+++ packages/plugins/plugin-crx/src/translations.ts
@@ -12,6 +12,15 @@ export const translations = [
       [meta.id]: {
         'plugin.name': 'CRX',
         'settings.title': 'Browser extension',
+        'toast.person.title': 'Added person',
+        'toast.organization.title': 'Added organization',
+        'toast.note.title': 'Added note',
+        'toast.error.title': 'Clip failed',
+        'toast.error.invalidPayload.message': 'Unrecognized clip payload.',
+        'toast.error.unsupportedVersion.message': 'Unsupported clip version — update Composer.',
+        'toast.error.unsupportedKind.message': 'Unsupported clip kind.',
+        'toast.error.noSpace.message': 'Open a space first.',
+        'toast.error.internal.message': 'Failed to save clip.',
       },
     },
   },
