+++ packages/plugins/plugin-crx/src/types/Clip.ts
