+++ packages/plugins/plugin-crx/src/types/capabilities.ts
@@ -12,9 +12,9 @@ import * as Settings from './Settings';
 
 export namespace CrxCapabilities {
   /**
-   * Writable atom holding the plugin's Settings. Other plugins (e.g.
-   * `@dxos/plugin-crx-bridge`) consume this to decide whether to materialize
-   * an incoming clip and how to surface it.
+   * Writable atom holding the plugin's Settings. Consumed by the clip
+   * listener to decide whether to materialize an incoming clip and how to
+   * surface it.
    */
   export const Settings = Capability.make<Atom.Writable<Settings.Settings>>(`${meta.id}.capability.settings`);
 }
