+++ packages/plugins/plugin-crx/vitest.config.ts
