+++ release-please-config.json
@@ -575,11 +575,6 @@
         },
         {
           "type": "json",
-          "path": "packages/plugins/plugin-crx-bridge/package.json",
-          "jsonpath": "$.version"
-        },
-        {
-          "type": "json",
           "path": "packages/plugins/plugin-daily-summary/package.json",
           "jsonpath": "$.version"
         },
